+++ ui/src/app/error/error.component.ts
@@ -9,7 +9,7 @@ import { AuthService } from '../services/auth.service';
   styleUrls: ['./error.component.less']
 })
 export class ErrorComponent implements OnInit {
-
+  showLoadingIndicator: any = false;
 
   constructor(public _route: Router, @Inject(MAT_DIALOG_DATA) public data: {
                   confirmText: string,
@@ -25,15 +25,24 @@ export class ErrorComponent implements OnInit {
   }
 
   public close(value: any) {
-    localStorage.clear();
-    this.authService.signOut().subscribe(()=>{
-      //localStorage.clear(); // clear all localstorage
-      this._route.navigate(["/auth/login"]);
+    this.showLoadingIndicator = true;
+    if(localStorage.length !== 0) {
+      localStorage.clear();
+      this.authService.signOut().subscribe(()=>{
+        //localStorage.clear(); // clear all localstorage
+        this.hideloader();
+        this.mdDialogRef.close(value);
+        this._route.navigate(["/auth/login"]);
+      }, (error) => {
+        this.hideloader();
+        this.mdDialogRef.close(value);
+        this._route.navigate(["/auth/login"]);
+      });
+    } else {
+      this.hideloader();
       this.mdDialogRef.close(value);
-    }, (error) => {
       this._route.navigate(["/auth/login"]);
-      this.mdDialogRef.close(value);
-    });
+    }
   }
 
   public confirm() {
@@ -47,6 +56,11 @@ export class ErrorComponent implements OnInit {
   }
 
   ngOnInit(): void {
+  }
+
+  hideloader() {
+    // Setting display of spinner
+    this.showLoadingIndicator = false;
   }
 
 }
